+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - Upcoming
+
+### Added
+-   macro: `month` to extract the month as an integer from a date
+-   macro: `year` to extract the year as an integer from a date
+
 ## [0.2.1] - 2025-08-12
 
 ### Fixed
@@ -30,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   initial commit
 
+[0.2.2]: https://github.com/linkFISH-Consulting/dbt-lf_utils/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/linkFISH-Consulting/dbt-lf_utils/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/linkFISH-Consulting/dbt-lf_utils/compare/v0.1.0...v0.2.0